refactor: Update Algolia component imports to match new file structure
