+++ overrides/app/components/algolia/dynamic-image/index.jsx
