+++ overrides/app/components/algolia/recommend/trendingFacets.jsx
