+++ overrides/app/components/algolia/recommend/utils/widgetheader.jsx
